Enhance GPG encryption/decryption API as Go common library
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/gauravaspire/gpg-gin-api?shareId=XXXX-XXXX-XXXX-XXXX).
