+++ README.md
@@ -7,8 +7,8 @@ This repository contains a **Go (Golang) web application** built with the **Gin 
 - [Features](#features)
 - [Prerequisites](#prerequisites)
 - [Installation](#installation)
-
 - [Usage](#usage)
+- [Library Usage](#library-usage)
 https://weaspire.atlassian.net/wiki/x/5wCMOg
 ---
 
@@ -53,5 +53,88 @@ Before you can run this project, ensure you have the following installed:
    ```bash
    git clone https://github.com/gauravaspire/gpg-gin-api.git
    cd gpg-gin-api
+   ```
 
+2. **Install dependencies:**
 
+   ```bash
+   go mod tidy
+   ```
+
+3. **Run the application:**
+
+   ```bash
+   go run main.go
+   ```
+
+---
+
+## Usage
+
+Once the application is running, you can use the following endpoints:
+
+1. **Encrypt Data:**
+
+   - **Endpoint:** `/encrypt`
+   - **Method:** `POST`
+   - **Description:** Encrypts the incoming data using the public key.
+   - **Request Body:** Raw data to be encrypted.
+   - **Response:** Encrypted data.
+
+   ```bash
+   curl -X POST http://localhost:8080/encrypt -d "Your data to encrypt"
+   ```
+
+2. **Decrypt Data:**
+
+   - **Endpoint:** `/decrypt`
+   - **Method:** `POST`
+   - **Description:** Decrypts the incoming encrypted data using the private key.
+   - **Request Body:** Raw encrypted data.
+   - **Response:** Decrypted data.
+
+   ```bash
+   curl -X POST http://localhost:8080/decrypt -d "Your encrypted data"
+   ```
+
+---
+
+## Library Usage
+
+The core GPG encryption and decryption logic has been refactored into a common library. You can use this library in your own Go projects.
+
+1. **Import the library:**
+
+   ```go
+   import "gpg-gin-api/library"
+   ```
+
+2. **Encrypt Data:**
+
+   ```go
+   publicKeyPath := "path/to/public.key"
+   data := []byte("Your data to encrypt")
+
+   encryptedData, err := library.Encrypt(publicKeyPath, data)
+   if err != nil {
+       // Handle error
+   }
+
+   // Use encryptedData
+   ```
+
+3. **Decrypt Data:**
+
+   ```go
+   privateKeyPath := "path/to/private.key"
+   encryptedData := []byte("Your encrypted data")
+
+   decryptedData, err := library.Decrypt(privateKeyPath, encryptedData)
+   if err != nil {
+       // Handle error
+   }
+
+   // Use decryptedData
+   ```
+
+---
